Drbn authentication (#1)
* Update index.php

Adds an admin icon and removes links if the person doesn't have authorizations to the module.

* Updates user / setUser

Updates user and setUser to redirect to the homepage if a user doesn't have privileges. Makes use of a new configuration.ini setting, createNewUserAutomatically, to determine if the traditional method of creating a read-only user is used.

* Add administrative options

Adds administrative options to the authentication settings.