Address additional numeric fields
Allow zeroes in resource cost fields: as with the taxRate field, other
numeric fields need to default to zero (0) rather than an empty string.

Format zero as currency in cost tables: Empty values in the cost table
are meant to be interpreted as a zero (0), formatted in the appropriate
locale, rather than an empty string.

Signed-off-by: Stephanie Leary <stephanie.leary@equinoxoli.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>